+++ cpp/shake_driver/inc/shake_io.h
@@ -31,7 +31,7 @@
 *
 *	This requires different methods of opening the connection, so this function
 *	acts as a standard "read" interface for the rest of the code */
-int read_bytes(shake_device_private* dev, char* buf, int bytes_to_read);
+size_t read_bytes(shake_device_private* dev, char* buf, size_t bytes_to_read);
 
 /*	The driver supports several possible methods of creating a connection. 
 *	It can be given a virtual serial port number (currently Windows only), a Bluetooth address
@@ -41,15 +41,15 @@ int read_bytes(shake_device_private* dev, char* buf, int bytes_to_read);
 *
 *	This requires different methods of opening the connection, so this function
 *	acts as a standard "write" interface for the rest of the code */
-int write_bytes(shake_device_private* dev, char* buf, int bytes_to_write);
+size_t write_bytes(shake_device_private* dev, char* buf, size_t bytes_to_write);
 
 /* the SHAKE often has problems dealing with large incoming data packets. this typically occurs with audio
 *	or vibration uploads which can total 1063 bytes per packet. To avoid overflowing the receive buffer on the
 *	SHAKE, this function is used to upload packets in chunks of <chunk_size> bytes, with a gap of <delay_ms>
 *	between successive chunks */
-int write_bytes_delayed(shake_device_private* dev, char* buf, int bytes_to_write, int chunk_size, int delay_ms);
+size_t write_bytes_delayed(shake_device_private* dev, char* buf, size_t bytes_to_write, size_t chunk_size, int delay_ms);
 
-int read_debug_bytes(shake_device_private* devpriv, char* buf, int bytes_to_read);
+size_t read_debug_bytes(shake_device_private* devpriv, char* buf, size_t bytes_to_read);
 
 #endif /* _SHAKE_IO_H_ */
